Store the list of active plugins that uses connection in an option (#15726)
* store the list of active plugins in an option

* prefix rename option

* tweaking order of hooks to test sync

* directly add option to sync whitelist

since the default whitelist is needed on WP.com it might not make sense
to use the filter in the package
but instead just update the default whitelist directly

* removes unused is_configured logic from Config

* Connection Package: Configuring the `Plugin_Storage` for unit tests.

Since the plugin storage configuration process has changes and has to be called directly, some unit tests started failing because of it.
The changes introduce an explicit call to the `configure()` method, so the storage would be properly configured for the tests to work properly.

* remove unused method

* fix sync whitelist options test

* store plugin slug and name

* fix logic to trigger refresh

Co-authored-by: Sergey Mitroshin <sergeymitr@gmail.com>

Committed via a GitHub action: https://github.com/automattic/jetpack/runs/110809600