Use the Password_Checker package (#18824)
* This is a skeleton PR for the Jetpack Password Checker package.

* Copying over the Jetpack_Password_Checker class.
Attempting to create a basic test, but it keeps failing.

* Adding a few mock functions to allow the phpunit tests to work.
Adding 2 basic password unit tests.

* Removing the Jetpack_ prefix from the class and file names.

* Switching over to WorDBless instead of brain monkey.
No need to create mock functions now.

* Fixing the broken phpunit tests.
I learned the hard way that the test class file name needs to be prefixed with "test-".

Fixing a bug in the get_other_user_data() method related to type checking.
Basically, the $user_data was expected to always contain an object.
When the user functions are mocked, the $user_data variable is set to false.
I just added some checking of the object vars before trying to use them.

* Check specifically for $this->user_id since that is what is actually needed to call get_userdata().
Return an empty array if $this->user_id is empty. Same if $user_data is false.

* Changing the isset() to empty(). The variable could be set with nonsense.

* Using the Password_Checker package.
Replacing the usage of jetpack_require_lib.

* Removing an errant namespace.

* Removing the password-checker package that was committed by mistake.

* Adding jetpack-password-checker to the composer.json file.
Changing authenticate_handler() to return a WP_Error() when the password test fails.

* Reverting this change so that authenticate_handler() always returns the $user.
Better to not change the behavior of this method for now.

* Removing the password-checker package from composer.json so that I can rebase from master.

* Adding changelog entry.

* Adding package to composer.json.

* [not verified] Revert "Adding package to composer.json."

This reverts commit cba8e0d88fdd9e2e8b137bcadad5ead69e1eb858.

* Adding the password-checker package to hopefully the correct composer.json.

* Adding changelog entry.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/686122182